+++ CHANGELOG.md
@@ -14,6 +14,7 @@ HumHub Changelog
 - Fix #7146: Fix search request by container guid
 - Fix #7141: Fix meta searching twice for the same keyword
 - Fix #7150: Remove js statement `with` to avoid error on build assets by grunt uglify
+- Fix #7156: Fix duplicated following spaces in the chooser widget
 
 1.16.1 (July 1, 2024)
 ---------------------